[APPC-3542] :ui:widgets modularization
