+++ app/build.gradle.kts
@@ -41,6 +41,7 @@ dependencies {
   implementation(project(":legacy:appcoinsRewards"))
   implementation(project(":legacy:skills"))
   implementation(project(":legacy:ethereumj-android"))
+  implementation(project(":ui:widgets"))
 
   implementation(libs.kotlin.coroutines)
