+++ app/src/main/res/layout/layout_transactions_activity.xml
@@ -30,7 +30,7 @@
           app:layout_constraintTop_toTopOf="parent"
           />
 
-      <com.asfoundation.wallet.widget.SystemView
+      <com.appcoins.wallet.widgets.SystemView
           android:id="@+id/system_view"
           android:layout_width="match_parent"
           android:layout_height="wrap_content"
