+++ app/src/main/res/layout/settings_promo_code_bottom_sheet_layout.xml
@@ -103,7 +103,7 @@
       />
 
 
-  <com.asfoundation.wallet.widget.SystemView
+  <com.appcoins.wallet.widgets.SystemView
       android:id="@+id/promo_code_bottom_sheet_system_view"
       android:layout_width="match_parent"
       android:layout_height="match_parent"
