+++ settings.gradle.kts
@@ -82,3 +82,4 @@ fun File.projectsAndModules(): Pair<Set<String>, Set<String>> {
   val formattedProjects = projects.map { it.replace(":", "/") }.toSet()
   return Pair(formattedProjects, modules)
 }
+include(":ui:widgets")
