+++ ui/widgets/build.gradle.kts
@@ -1,0 +1,23 @@
+plugins {
+  id("appcoins.android.library")
+}
+
+android {
+  namespace = "com.appcoins.wallet.widgets"
+  buildTypes {
+    release {
+    }
+    debug {
+    }
+  }
+}
+
+dependencies {
+  implementation(project(":ui:commons"))
+
+  implementation(libs.androidx.recyclerview)
+  implementation(libs.androidx.navigation.ui)
+
+  implementation(libs.zxing.android)
+  implementation(libs.glide)
+}
